+++ index.html
@@ -6,8 +6,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="/src/css/style.css">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js" defer></script><script src="/src/js/post.js" defer></script>
     <script src="/src/js/script.js" defer></script>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js" defer></script>
     <title>Document</title>
 </head>
 <body>
@@ -22,20 +22,6 @@
         </div>
       </nav>
 
-
-
-    <div class="card text-center postblog">
-        <div class="card-header">
-          usuario  y mail user
-        </div>
-        <div class="card-body">
-          <h5 class="card-title">titulo</h5>
-          <p class="card-text">cuerpo del post</p>
-          <!-- <a href="#" class="btn btn-primary">coments</a> -->
-        </div>
-      </div>
-
-
-      
+    <div id="redyblog"></div>
 </body>
 </html>
