+++ jsconfig.json
@@ -1,0 +1,8 @@
+{
+    "typeAcquisition": {
+        "include": [
+            "jquery"
+        ]
+    }
+
+}
